Perform and plot a Principal Components Analysis

Perform and plot a Principal Components Analysis

Usage

plot_pca(
  log_counts,
  sample_metadata,
  samples_to_include,
  samples_to_rename_manually,
  group_column,
  label_column,
  color_values,
  principal_component_on_x_axis = 1,
  principal_component_on_y_axis = 2,
  legend_position_for_pca = "top",
  point_size_for_pca = 1,
  add_label_to_pca = TRUE,
  label_font_size = 3,
  label_offset_y_ = 2,
  label_offset_x_ = 2
)

Arguments

log_counts

log-transformed filtered counts

samples_to_include

samples in sample_metadata to include in the analysis

samples_to_rename_manually

If you do not have a Plot Labels Column in your sample metadata table, you can use this parameter to rename samples manually for display on the PCA plot. Use "Add item" to add each additional sample for renaming. Use the following format to describe which old name (in your sample metadata table) you want to rename to which new name: old_name: new_name

group_column

The column from your input Sample Metadata table containing the sample group information. This is usually a column showing to which experimental treatments each sample belongs (e.g. WildType, Knockout, Tumor, Normal, Before, After, etc.). Only columns of Text type from your input Sample Metadata will be available to select for this parameter.

label_column

The column from your input Sample Metadata table containing the sample labels as you wish them to appear in the plots produced by this template. This can be the same Sample Names Column. However, you may desire different labels to display on your figure (e.g. shorter labels are sometimes preferred on plots). In that case, select the column with your preferred Labels here. The selected column should contain unique names for each sample.

principal_component_on_x_axis

The principle component to plot on the x-axis for the PCA plot. Choices include 1, 2, 3, ... (default: 1)

principal_component_on_y_axis

The principle component to plot on the y-axis for the PCA plot. Choices include 1, 2, 3, ... (default: 2)

legend_position_for_pca

legend position for the PCA plot

point_size_for_pca

geom point size for the PCA plot

add_label_to_pca

label points on the PCA plot

label_font_size

label font size for the PCA plot

label_offset_y_

label offset y for the PCA plot

label_offset_x_

label offset x for the PCA plot

Value

PCA plot