Run DESeq2 on a reneeDataSet

Run DESeq2 on a reneeDataSet

Usage

run_deseq2(renee_ds, design, ...)

Arguments

renee_ds

reneeDataSet object

design

model formula for experimental design. Columns must exist in meta_dat.

...

remaining variables are forwarded to DESeq2::DESeq().

Value

reneeDataSet object with DESeq2 slot filled

Examples

renee_ds <- create_reneeDataSet_from_files(
  system.file("extdata",
    "RSEM.genes.expected_count.all_samples.txt",
    package = "reneeTools"
  ),
  system.file("extdata", "sample_metadata.tsv",
    package = "reneeTools"
  )
)
#> Rows: 58929 Columns: 6
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: "\t"
#> chr (2): gene_id, GeneName
#> dbl (4): KO_S3, KO_S4, WT_S1, WT_S2
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Rows: 4 Columns: 2
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: "\t"
#> chr (2): sample_id, condition
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
renee_ds <- run_deseq2(renee_ds, ~condition)
#> Warning: some variables in design formula are characters, converting to factors
#> estimating size factors
#> estimating dispersions
#> gene-wise dispersion estimates
#> mean-dispersion relationship
#> -- note: fitType='parametric', but the dispersion trend was not well captured by the
#>    function: y = a/x + b, and a local regression fit was automatically substituted.
#>    specify fitType='local' or 'mean' to avoid this message next time.
#> final dispersion estimates
#> fitting model and testing