Run DESeq2 on a reneeDataSet

Run DESeq2 on a reneeDataSet

Usage

run_deseq2(renee_ds, design, ...)

Arguments

renee_ds

reneeDataSet object

design

model formula for experimental design. Columns must exist in meta_dat.

...

remaining variables are forwarded to DESeq2::DESeq().

Value

reneeDataSet object with DESeq2 slot filled

Examples

rds <- reneeDataSetFromFiles(
  system.file("extdata",
    "RSEM.genes.expected_count.all_samples.txt",
    package = "reneeTools"
  ),
  system.file("extdata", "sample_metadata.tsv",
    package = "reneeTools"
  )
)
#> Rows: 58929 Columns: 6
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: "\t"
#> chr (2): gene_id, GeneName
#> dbl (4): KO_S3, KO_S4, WT_S1, WT_S2
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Rows: 4 Columns: 2
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: "\t"
#> chr (2): sample_id, condition
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
run_deseq2(rds, ~condition)
#> Warning: some variables in design formula are characters, converting to factors
#> estimating size factors
#> estimating dispersions
#> gene-wise dispersion estimates
#> mean-dispersion relationship
#> -- note: fitType='parametric', but the dispersion trend was not well captured by the
#>    function: y = a/x + b, and a local regression fit was automatically substituted.
#>    specify fitType='local' or 'mean' to avoid this message next time.
#> final dispersion estimates
#> fitting model and testing
#> <renee>
#>  @ counts     : int [1:58929, 1:4] 0 0 0 0 0 0 0 0 0 0 ...
#>  .. - attr(*, "dimnames")=List of 2
#>  ..  ..$ : chr [1:58929] "ENSG00000121410.11" "ENSG00000268895.5" "ENSG00000148584.15" "ENSG00000175899.14" ...
#>  ..  ..$ : chr [1:4] "KO_S3" "KO_S4" "WT_S1" "WT_S2"
#>  @ sample_meta:'data.frame':	4 obs. of  1 variable:
#>  .. $ condition: chr  "knockout" "knockout" "wildtype" "wildtype"
#>  @ analyses   :List of 1
#>  .. $ deseq2_ds:Formal class 'DESeqDataSet' [package "DESeq2"] with 8 slots
#>  ..  .. ..@ design            :Class 'formula'  language ~condition
#>  ..  .. .. .. ..- attr(*, ".Environment")=<environment: 0x559716014a38> 
#>  ..  .. ..@ dispersionFunction:function (means)  
#>  ..  .. .. ..- attr(*, "fitType")= chr "local"
#>  ..  .. .. ..- attr(*, "varLogDispEsts")= num 0.0119
#>  ..  .. .. ..- attr(*, "dispPriorVar")= num 0.25
#>  ..  .. ..@ rowRanges         :Formal class 'CompressedGRangesList' [package "GenomicRanges"] with 5 slots
#>  ..  .. .. .. ..@ unlistData     :Formal class 'GRanges' [package "GenomicRanges"] with 7 slots
#>  ..  .. .. .. .. .. ..@ seqnames       :Formal class 'Rle' [package "S4Vectors"] with 4 slots
#>  ..  .. .. .. .. .. .. .. ..@ values         : Factor w/ 0 levels: 
#>  ..  .. .. .. .. .. .. .. ..@ lengths        : int(0) 
#>  ..  .. .. .. .. .. .. .. ..@ elementMetadata: NULL
#>  ..  .. .. .. .. .. .. .. ..@ metadata       : list()
#>  ..  .. .. .. .. .. ..@ ranges         :Formal class 'IRanges' [package "IRanges"] with 6 slots
#>  ..  .. .. .. .. .. .. .. ..@ start          : int(0) 
#>  ..  .. .. .. .. .. .. .. ..@ width          : int(0) 
#>  ..  .. .. .. .. .. .. .. ..@ NAMES          : NULL
#>  ..  .. .. .. .. .. .. .. ..@ elementType    : chr "ANY"
#>  ..  .. .. .. .. .. .. .. ..@ elementMetadata: NULL
#>  ..  .. .. .. .. .. .. .. ..@ metadata       : list()
#>  ..  .. .. .. .. .. ..@ strand         :Formal class 'Rle' [package "S4Vectors"] with 4 slots
#>  ..  .. .. .. .. .. .. .. ..@ values         : Factor w/ 3 levels "+","-","*": 
#>  ..  .. .. .. .. .. .. .. ..@ lengths        : int(0) 
#>  ..  .. .. .. .. .. .. .. ..@ elementMetadata: NULL
#>  ..  .. .. .. .. .. .. .. ..@ metadata       : list()
#>  ..  .. .. .. .. .. ..@ seqinfo        :Formal class 'Seqinfo' [package "GenomeInfoDb"] with 4 slots
#>  ..  .. .. .. .. .. .. .. ..@ seqnames   : chr(0) 
#>  ..  .. .. .. .. .. .. .. ..@ seqlengths : int(0) 
#>  ..  .. .. .. .. .. .. .. ..@ is_circular: logi(0) 
#>  ..  .. .. .. .. .. .. .. ..@ genome     : chr(0) 
#>  ..  .. .. .. .. .. ..@ elementMetadata:Formal class 'DFrame' [package "S4Vectors"] with 6 slots
#>  ..  .. .. .. .. .. .. .. ..@ rownames       : NULL
#>  ..  .. .. .. .. .. .. .. ..@ nrows          : int 0
#>  ..  .. .. .. .. .. .. .. ..@ elementType    : chr "ANY"
#>  ..  .. .. .. .. .. .. .. ..@ elementMetadata: NULL
#>  ..  .. .. .. .. .. .. .. ..@ metadata       : list()
#>  ..  .. .. .. .. .. .. .. ..@ listData       : Named list()
#>  ..  .. .. .. .. .. ..@ elementType    : chr "ANY"
#>  ..  .. .. .. .. .. ..@ metadata       : list()
#>  ..  .. .. .. ..@ elementMetadata:Formal class 'DFrame' [package "S4Vectors"] with 6 slots
#>  ..  .. .. .. .. .. ..@ rownames       : NULL
#>  ..  .. .. .. .. .. ..@ nrows          : int 58929
#>  ..  .. .. .. .. .. ..@ elementType    : chr "ANY"
#>  ..  .. .. .. .. .. ..@ elementMetadata:Formal class 'DFrame' [package "S4Vectors"] with 6 slots
#>  ..  .. .. .. .. .. .. .. ..@ rownames       : NULL
#>  ..  .. .. .. .. .. .. .. ..@ nrows          : int 22
#>  ..  .. .. .. .. .. .. .. ..@ elementType    : chr "ANY"
#>  ..  .. .. .. .. .. .. .. ..@ elementMetadata: NULL
#>  ..  .. .. .. .. .. .. .. ..@ metadata       : list()
#>  ..  .. .. .. .. .. .. .. ..@ listData       :List of 2
#>  ..  .. .. .. .. .. .. .. .. ..$ type       : chr [1:22] "intermediate" "intermediate" "intermediate" "intermediate" ...
#>  ..  .. .. .. .. .. .. .. .. ..$ description: chr [1:22] "mean of normalized counts for all samples" "variance of normalized counts for all samples" "all counts for a gene are zero" "gene-wise estimates of dispersion" ...
#>  ..  .. .. .. .. .. ..@ metadata       : list()
#>  ..  .. .. .. .. .. ..@ listData       :List of 22
#>  ..  .. .. .. .. .. .. ..$ baseMean                                    : num [1:58929] 0 0 0 0 0 0 0 0 0 0 ...
#>  ..  .. .. .. .. .. .. ..$ baseVar                                     : num [1:58929] 0 0 0 0 0 0 0 0 0 0 ...
#>  ..  .. .. .. .. .. .. ..$ allZero                                     : logi [1:58929] TRUE TRUE TRUE TRUE TRUE TRUE ...
#>  ..  .. .. .. .. .. .. ..$ dispGeneEst                                 : num [1:58929] NA NA NA NA NA NA NA NA NA NA ...
#>  ..  .. .. .. .. .. .. ..$ dispGeneIter                                : num [1:58929] NA NA NA NA NA NA NA NA NA NA ...
#>  ..  .. .. .. .. .. .. ..$ dispFit                                     : num [1:58929] NA NA NA NA NA NA NA NA NA NA ...
#>  ..  .. .. .. .. .. .. ..$ dispersion                                  : num [1:58929] NA NA NA NA NA NA NA NA NA NA ...
#>  ..  .. .. .. .. .. .. ..$ dispIter                                    : num [1:58929] NA NA NA NA NA NA NA NA NA NA ...
#>  ..  .. .. .. .. .. .. ..$ dispOutlier                                 : logi [1:58929] NA NA NA NA NA NA ...
#>  ..  .. .. .. .. .. .. ..$ dispMAP                                     : num [1:58929] NA NA NA NA NA NA NA NA NA NA ...
#>  ..  .. .. .. .. .. .. ..$ Intercept                                   : num [1:58929] NA NA NA NA NA NA NA NA NA NA ...
#>  ..  .. .. .. .. .. .. ..$ condition_wildtype_vs_knockout              : num [1:58929] NA NA NA NA NA NA NA NA NA NA ...
#>  ..  .. .. .. .. .. .. ..$ SE_Intercept                                : num [1:58929] NA NA NA NA NA NA NA NA NA NA ...
#>  ..  .. .. .. .. .. .. ..$ SE_condition_wildtype_vs_knockout           : num [1:58929] NA NA NA NA NA NA NA NA NA NA ...
#>  ..  .. .. .. .. .. .. ..$ WaldStatistic_Intercept                     : num [1:58929] NA NA NA NA NA NA NA NA NA NA ...
#>  ..  .. .. .. .. .. .. ..$ WaldStatistic_condition_wildtype_vs_knockout: num [1:58929] NA NA NA NA NA NA NA NA NA NA ...
#>  ..  .. .. .. .. .. .. ..$ WaldPvalue_Intercept                        : num [1:58929] NA NA NA NA NA NA NA NA NA NA ...
#>  ..  .. .. .. .. .. .. ..$ WaldPvalue_condition_wildtype_vs_knockout   : num [1:58929] NA NA NA NA NA NA NA NA NA NA ...
#>  ..  .. .. .. .. .. .. ..$ betaConv                                    : logi [1:58929] NA NA NA NA NA NA ...
#>  ..  .. .. .. .. .. .. ..$ betaIter                                    : num [1:58929] NA NA NA NA NA NA NA NA NA NA ...
#>  ..  .. .. .. .. .. .. ..$ deviance                                    : num [1:58929] NA NA NA NA NA NA NA NA NA NA ...
#>  ..  .. .. .. .. .. .. ..$ maxCooks                                    : logi [1:58929] NA NA NA NA NA NA ...
#>  ..  .. .. .. ..@ elementType    : chr "GRanges"
#>  ..  .. .. .. ..@ metadata       : list()
#>  ..  .. .. .. ..@ partitioning   :Formal class 'PartitioningByEnd' [package "IRanges"] with 5 slots
#>  ..  .. .. .. .. .. ..@ end            : int [1:58929] 0 0 0 0 0 0 0 0 0 0 ...
#>  ..  .. .. .. .. .. ..@ NAMES          : chr [1:58929] "ENSG00000121410.11" "ENSG00000268895.5" "ENSG00000148584.15" "ENSG00000175899.14" ...
#>  ..  .. .. .. .. .. ..@ elementType    : chr "ANY"
#>  ..  .. .. .. .. .. ..@ elementMetadata: NULL
#>  ..  .. .. .. .. .. ..@ metadata       : list()
#>  ..  .. ..@ colData           :Formal class 'DFrame' [package "S4Vectors"] with 6 slots
#>  ..  .. .. .. ..@ rownames       : chr [1:4] "KO_S3" "KO_S4" "WT_S1" "WT_S2"
#>  ..  .. .. .. ..@ nrows          : int 4
#>  ..  .. .. .. ..@ elementType    : chr "ANY"
#>  ..  .. .. .. ..@ elementMetadata:Formal class 'DFrame' [package "S4Vectors"] with 6 slots
#>  ..  .. .. .. .. .. ..@ rownames       : NULL
#>  ..  .. .. .. .. .. ..@ nrows          : int 2
#>  ..  .. .. .. .. .. ..@ elementType    : chr "ANY"
#>  ..  .. .. .. .. .. ..@ elementMetadata: NULL
#>  ..  .. .. .. .. .. ..@ metadata       : list()
#>  ..  .. .. .. .. .. ..@ listData       :List of 2
#>  ..  .. .. .. .. .. .. ..$ type       : chr [1:2] "input" "intermediate"
#>  ..  .. .. .. .. .. .. ..$ description: chr [1:2] "" "a scaling factor for columns"
#>  ..  .. .. .. ..@ metadata       : list()
#>  ..  .. .. .. ..@ listData       :List of 2
#>  ..  .. .. .. .. ..$ condition : Factor w/ 2 levels "knockout","wildtype": 1 1 2 2
#>  ..  .. .. .. .. ..$ sizeFactor: Named num [1:4] 0.76 0.719 1.245 1.682
#>  ..  .. .. .. .. .. ..- attr(*, "names")= chr [1:4] "KO_S3" "KO_S4" "WT_S1" "WT_S2"
#>  ..  .. ..@ assays            :Formal class 'SimpleAssays' [package "SummarizedExperiment"] with 1 slot
#>  ..  .. .. .. ..@ data:Formal class 'SimpleList' [package "S4Vectors"] with 4 slots
#>  ..  .. .. .. .. .. ..@ listData       :List of 4
#>  ..  .. .. .. .. .. .. ..$ counts: int [1:58929, 1:4] 0 0 0 0 0 0 0 0 0 0 ...
#>  ..  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
#>  ..  .. .. .. .. .. .. .. .. ..$ : chr [1:58929] "ENSG00000121410.11" "ENSG00000268895.5" "ENSG00000148584.15" "ENSG00000175899.14" ...
#>  ..  .. .. .. .. .. .. .. .. ..$ : chr [1:4] "KO_S3" "KO_S4" "WT_S1" "WT_S2"
#>  ..  .. .. .. .. .. .. ..$ mu    : num [1:58929, 1:4] NA NA NA NA NA NA NA NA NA NA ...
#>  ..  .. .. .. .. .. .. ..$ H     : num [1:58929, 1:4] NA NA NA NA NA NA NA NA NA NA ...
#>  ..  .. .. .. .. .. .. ..$ cooks : num [1:58929, 1:4] NA NA NA NA NA NA NA NA NA NA ...
#>  ..  .. .. .. .. .. ..@ elementType    : chr "ANY"
#>  ..  .. .. .. .. .. ..@ elementMetadata: NULL
#>  ..  .. .. .. .. .. ..@ metadata       : list()
#>  ..  .. ..@ NAMES             : NULL
#>  ..  .. ..@ elementMetadata   :Formal class 'DFrame' [package "S4Vectors"] with 6 slots
#>  ..  .. .. .. ..@ rownames       : NULL
#>  ..  .. .. .. ..@ nrows          : int 58929
#>  ..  .. .. .. ..@ elementType    : chr "ANY"
#>  ..  .. .. .. ..@ elementMetadata: NULL
#>  ..  .. .. .. ..@ metadata       : list()
#>  ..  .. .. .. ..@ listData       : Named list()
#>  ..  .. ..@ metadata          :List of 1
#>  ..  .. .. ..$ version:Classes 'package_version', 'numeric_version'  hidden list of 1
#>  ..  .. .. .. ..$ : int [1:3] 1 42 0
#>  ..  .. ..$ betaPrior      : logi FALSE
#>  ..  .. ..$ modelMatrixType: chr "standard"
#>  ..  .. ..$ betaPriorVar   : num [1:2] 1e+06 1e+06
#>  ..  .. ..$ modelMatrix    : num [1:4, 1:2] 1 1 1 1 0 0 1 1
#>  ..  .. .. ..- attr(*, "dimnames")=List of 2
#>  ..  .. .. .. ..$ : chr [1:4] "KO_S3" "KO_S4" "WT_S1" "WT_S2"
#>  ..  .. .. .. ..$ : chr [1:2] "Intercept" "condition_wildtype_vs_knockout"
#>  ..  .. .. ..- attr(*, "assign")= int [1:2] 0 1
#>  ..  .. .. ..- attr(*, "contrasts")=List of 1
#>  ..  .. .. .. ..$ condition: chr "contr.treatment"
#>  ..  .. ..$ test           : chr "Wald"
#>  ..  .. ..$ dispModelMatrix: num [1:4, 1:2] 1 1 1 1 0 0 1 1
#>  ..  .. .. ..- attr(*, "dimnames")=List of 2
#>  ..  .. .. .. ..$ : chr [1:4] "KO_S3" "KO_S4" "WT_S1" "WT_S2"
#>  ..  .. .. .. ..$ : chr [1:2] "(Intercept)" "conditionwildtype"
#>  ..  .. .. ..- attr(*, "assign")= int [1:2] 0 1
#>  ..  .. .. ..- attr(*, "contrasts")=List of 1
#>  ..  .. .. .. ..$ condition: chr "contr.treatment"